Added the beginnings of a search bar in main.html. See full mesg.
Basically I'm not able to correctly encode the search terms using
encodeURI, and that means that it's not going to play nice with special
characters. So, I need to figure out exactly why that's happenting.

Also note that we will need to add this code to all the html files,
including the added jquery at the bottom. I need to do some research to
see if we can use HTML includes, since it's very frustrating to have all
this duplicated code.