Fix CSS sourcing and note display.
Before, the CSS file paths were relative paths, not absolute paths. So
some pages weren't using the same css as the others, which was looking
less than perfect.

I also changed the /notes page so that it would display all of the
notes, rather than just the first line. This is better UI until I get it
to the point where each note title is a link.